+++ src/Resources/contao/dca/tl_page.php
@@ -1,3 +1,0 @@
-<?php
-
-\Contao\System::getContainer()->get('huh.utils.cache.database_tree')->registerDcaToCacheTree('tl_page', ['type = ?'], ['root'], ['order' => 'sorting']);